Dependency and submodule changes for nodejs 0.12+.
Because the ninjablocks repos (and related ones of former Ninjablocks
staff) appear essentially unmaintained, many of these updates unfortunately
require changing the dependency to a fork that contains the necessary fixes.

In the course of the submodule updates mandated by nodejs 0.12+, also
updated others to include earlier improvements and fixes were made
after the last update to the registered submodule commits.